added a new pathology section to summary metadata
